Don't automatically re-sort collections
When using `SortedCollection` or `SortedList` to store objects, calling `map()` on the list with a callback that returns a different object will cause unwanted order in the lists due to the sort function potentially checking for objects. To avoid this, the `map()` function creates a new instance of itself without sorting the elements in the constructor.

Additionally, the same behavior is used in the `filter()` and `partition()` methods since we can assume that the list has been previously sorted and does not need to be sorted again. This increases performance when applying a callback to a large list.